requirements.txt
openai
requests
beautifulsoup4
pytz
python-telegram-bot
google-auth
google-api-python-client
